+++ astropy/io/fits/connect.py
@@ -11,8 +11,9 @@ from astropy import units as u
 from astropy.table import Table, serialize, meta, Column, MaskedColumn
 from astropy.table.table import has_info_class
 from astropy.time import Time
-from astropy.utils.exceptions import AstropyUserWarning
 from astropy.utils.data_info import MixinInfo, serialize_context_as
+from astropy.utils.exceptions import (AstropyUserWarning,
+                                      AstropyDeprecationWarning)
 from . import HDUList, TableHDU, BinTableHDU, GroupsHDU
 from .column import KEYWORD_NAMES, _fortran_to_python_format
 from .convenience import table_to_hdu
@@ -158,7 +159,7 @@ def read_table_fits(input, hdu=None, astropy_native=False, memmap=False,
     if isinstance(input, HDUList):
 
         # Parse all table objects
-        tables = OrderedDict()
+        tables = dict()
         for ihdu, hdu_item in enumerate(input):
             if isinstance(hdu_item, (TableHDU, BinTableHDU, GroupsHDU)):
                 tables[ihdu] = hdu_item
@@ -167,7 +168,7 @@ def read_table_fits(input, hdu=None, astropy_native=False, memmap=False,
             if hdu is None:
                 warnings.warn("hdu= was not specified but multiple tables"
                               " are present, reading in first available"
-                              " table (hdu={})".format(first(tables)),
+                              f" table (hdu={first(tables)})",
                               AstropyUserWarning)
                 hdu = first(tables)
 
@@ -181,7 +182,24 @@ def read_table_fits(input, hdu=None, astropy_native=False, memmap=False,
                 raise ValueError(f"No table found in hdu={hdu}")
 
         elif len(tables) == 1:
+            if hdu is not None:
+                msg = None
+                try:
+                    hdi = input.index_of(hdu)
+                except KeyError:
+                    msg = f"Specified hdu={hdu} not found"
+                else:
+                    if hdi >= len(input):
+                        msg = f"Specified hdu={hdu} not found"
+                    elif hdi not in tables:
+                        msg = f"No table found in specified hdu={hdu}"
+                if msg is not None:
+                    warnings.warn(f"{msg}, reading in first available table "
+                                  f"(hdu={first(tables)}) instead. This will"
+                                  " result in an error in future versions!",
+                                  AstropyDeprecationWarning)
             table = tables[first(tables)]
+
         else:
             raise ValueError("No table found")